Only remove old directory if it already exists, and only install bundler if it isn't already installed

+++ download_sourcery.sh
@@ -1,9 +1,15 @@
+#!/bin/bash
+
+if !(gem query -i -n bundler); then
 echo "Installing any dependencies..."
 gem install bundler
 bundle install
+fi
 
+SOURCERY_DIR="./Sourcery"
+if [ -d "$SOURCERY_DIR" ] ; then
 echo "Removing old Sourcery directory..."
 rm -rf ./Sourcery
+fi
 
-echo "Cloning Sourcery..."
-bundle exec git clone https://github.com/krzysztofzablocki/Sourcery
+bundle exec git clone https://github.com/krzysztofzablocki/Sourcery $SOURCERY_DIR
